Load Boot's REPL middleware when starting nREPL.
The same middleware and dependencies that would be loaded for the Boot
`repl` task are also used for the nREPL hosted by the HTTP server.  This
includes the middleware configured in a user's `boot.profile` (assuming
that they've followed the standard recipe of concating additional
dependencies to `boot.repl/*default-dependencies*`;
cf. https://github.com/boot-clj/boot/wiki/Cider-REPL#a-better-way)

It is particularly useful for CIDER users, who require `cider-nrepl` and
associated middleware to be loaded into the nREPL server for CIDER to be
fully functional.